refactor : background dto구조 변경
