+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Background/entity/ImageTransform.java
@@ -1,4 +1,4 @@
-package com.techeerpicture.TecheerPicture.Background;
+package com.techeerpicture.TecheerPicture.Background.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
